[Design] #5 - TicketView 생성
